refactor of pattern and justification grammar
